Clean up and move things around
